Split fixtures into caching and signal generator
